Switching to MariaDB
The vector store can now save vectors with up to 65532 dimensions. https://mariadb.com/docs/server/reference/data-types/numeric-data-types/vector

 - PostgreSQL is replaced with MariaDB
 - The VectorStoreExtension has been implemented for MariaDB. The implementation has been copied from the PgVectorStoreExtension, since they are very similar.
 - The table name is set based on the given model. Here only the property names were changed.
 - application YAMLs have been updated
