+++ docker/docker-compose.yaml
@@ -31,17 +31,20 @@ services:
       garage:
         condition: service_started
   db:
-    image: pgvector/pgvector:pg17
+    image: mariadb:11.8.2-ubi9
     ports:
-      - "5432:5432"
+      - "3306:3306"
     environment:
-      - POSTGRES_USER=example
-      - POSTGRES_PASSWORD=example
-      - POSTGRES_DB=example
-      - PGUSER=example
+      - MARIADB_USER=example
+      - MARIADB_PASSWORD=example
+      - MARIADB_DATABASE=example
+      - MARIADB_ROOT_PASSWORD=example
     healthcheck:
-      test: [ "CMD-SHELL", "pg_isready" ]
-      start_period: 5s
+      test: [ "CMD", "healthcheck.sh", "--connect", "--innodb_initialized" ]
+      start_period: 10s
+      interval: 10s
+      timeout: 5s
+      retries: 3
 
   garage:
     image: dxflrs/garage:v2.0.0
