+++ pom.xml
@@ -65,7 +65,7 @@
         </dependency>
         <dependency>
             <groupId>org.springframework.ai</groupId>
-            <artifactId>spring-ai-starter-vector-store-pgvector</artifactId>
+            <artifactId>spring-ai-starter-vector-store-mariadb</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.ai</groupId>
